refactor: Extract shared API dependencies (Issue #164)
- Create api/dependencies.py with shared get_index() and set_index()
- Remove duplicated code from navigation.py, content.py, manipulation.py
- Simplify app.py to use single dependencies.set_index() call
- ~40 lines of duplicated code removed

All 435 tests pass, no behavior changes.
