+++ src/dacli/api/navigation.py
@@ -8,8 +8,8 @@ Provides endpoints for navigating the document structure:
 
 from fastapi import APIRouter, HTTPException, Path, Query
 
+from dacli.api.dependencies import get_index
 from dacli.api.models import (
-    ErrorDetail,
     ErrorResponse,
     LocationResponse,
     SectionDetailResponse,
@@ -18,33 +18,8 @@ from dacli.api.models import (
     SectionSummary,
     StructureResponse,
 )
-from dacli.structure_index import StructureIndex
 
 router = APIRouter(prefix="/api/v1", tags=["Navigation"])
-
-# Global index reference - will be set by create_app
-_index: StructureIndex | None = None
-
-
-def set_index(index: StructureIndex) -> None:
-    """Set the global structure index."""
-    global _index
-    _index = index
-
-
-def get_index() -> StructureIndex:
-    """Get the global structure index."""
-    if _index is None:
-        raise HTTPException(
-            status_code=503,
-            detail=ErrorResponse(
-                error=ErrorDetail(
-                    code="INDEX_NOT_READY",
-                    message="Server index is not initialized",
-                )
-            ).model_dump(),
-        )
-    return _index
 
 
 @router.get(
